+++ deeplc/__main__.py
@@ -7,9 +7,9 @@ __maintainer__ = ["Robbin Bouwmeester", "Ralf Gabriels"]
 __email__ = ["Robbin.Bouwmeester@ugent.be", "Ralf.Gabriels@ugent.be"]
 
 import logging
-import multiprocessing
 import os
 import sys
+import warnings
 
 import pandas as pd
 from matplotlib import pyplot as plt
@@ -50,84 +50,44 @@ def main(gui=False):
     argu = parse_arguments(gui=gui)
 
     setup_logging(argu.log_level)
-    if argu.log_level.lower() != "debug":
-        logging.getLogger('tensorflow').setLevel(logging.CRITICAL)
-        os.environ['KMP_WARNINGS'] = '0'
-        os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 
-    max_threads = multiprocessing.cpu_count()
-    if not argu.n_threads:
-        argu.n_threads = max_threads
-    elif argu.n_threads > max_threads:
-        argu.n_threads = max_threads
+    # Reset logging levels if DEBUG (see deeplc.py)
+    if argu.log_level.lower() == "debug":
+        os.environ['TF_CPP_MIN_LOG_LEVEL'] = '0'
+        logging.getLogger('tensorflow').setLevel(logging.DEBUG)
+        warnings.filterwarnings('default', category=DeprecationWarning)
+        warnings.filterwarnings('default', category=FutureWarning)
+        warnings.filterwarnings('default', category=UserWarning)
+    else:
+        os.environ['KMP_WARNINGS'] = '0'
 
     try:
-        run(file_pred=argu.file_pred,
-            file_cal=argu.file_cal,
-            file_pred_out=argu.file_pred_out,
-            file_model=argu.file_model,
-            n_threads=argu.n_threads,
-            verbose=True,
-            split_cal=argu.split_cal,
-            dict_divider=argu.dict_divider,
-            batch_num=argu.batch_num,
-            plot_predictions=argu.plot_predictions,
-            write_library=argu.write_library,
-            use_library=argu.use_library
-            )
+        run(**vars(argu))
     except DeepLCError as e:
         logger.exception(e)
         sys.exit(1)
 
 
-def run(file_pred,
-        file_cal=None,
-        file_pred_out=None,
-        file_model=None,
-        n_threads=None,
-        verbose=False,
-        split_cal=50,
-        dict_divider=50,
-        batch_num=50000,
-        pygam_calibration=False,
-        plot_predictions=False,
-        write_library=False,
-        use_library=""
-        ):
-    """
-    Main function to run the DeepLC code
-
-    Parameters
-    ----------
-    file_pred : str
-        the file in peprec format that we need to make predictions for
-        this file is not required to contain a tr column
-    file_cal : str
-        the file in peprec format that we use for calibrating the prediction
-        model. This file is required to contain a tr column
-    file_pred_out : str
-        outfile for predictions, the file is in peprec format and predictions
-        are added in the column TODO
-    file_model : str | list | None
-        the model(s) to try for retention time prediction can be a single
-        location or several locations for multiple models to try
-    n_threads : int
-        number of threads to run mainly the feature extraction on
-    split_cal : int
-        number of splits or divisions to use for the calibration
-    dict_divider : int
-        TODO
-    batch_num : int
-        TODO
-    plot_predictions : bool
-        Save scatter plot of predictions vs observations
-
-    Returns
-    -------
-    None
-    """
+def run(
+    file_pred,
+    file_cal=None,
+    file_pred_out=None,
+    plot_predictions=False,
+    file_model=None,
+    pygam_calibration=False,
+    split_cal=50,
+    dict_divider=50,
+    use_library=None,
+    write_library=False,
+    batch_num=50000,
+    n_threads=None,
+    log_level="info",
+    verbose=True,
+):
+    """Run DeepLC."""
 
     logger.info("Using DeepLC version %s", __version__)
+    logger.debug("Using %i CPU threads", n_threads)
 
     if not file_cal and file_model != None:
         fm_dict = {}
@@ -158,27 +118,31 @@ def run(file_pred,
     # use the default settings for DeepLC. Here we want to use a model that does
     # not use RDKit features so we skip the chemical descriptor making
     # procedure.
-    f_extractor = FeatExtractor(add_sum_feat=False,
-                                ptm_add_feat=False,
-                                ptm_subtract_feat=False,
-                                standard_feat=False,
-                                chem_descr_feat=False,
-                                add_comp_feat=False,
-                                cnn_feats=True,
-                                verbose=verbose)
+    f_extractor = FeatExtractor(
+        add_sum_feat=False,
+        ptm_add_feat=False,
+        ptm_subtract_feat=False,
+        standard_feat=False,
+        chem_descr_feat=False,
+        add_comp_feat=False,
+        cnn_feats=True,
+        verbose=verbose
+    )
 
-    # Make the DeepLC object that will handle making predictions and
-    # calibration
-    dlc = DeepLC(path_model=file_model,
-                 f_extractor=f_extractor,
-                 split_cal=split_cal,
-                 pygam_calibration=pygam_calibration,
-                 cnn_model=True,
-                 n_jobs=n_threads,
-                 verbose=verbose,
-                 batch_num=batch_num,
-                 write_library=write_library,
-                 use_library=use_library)
+    # Make the DeepLC object that will handle making predictions and calibration
+    dlc = DeepLC(
+        path_model=file_model,
+        f_extractor=f_extractor,
+        cnn_model=True,
+        pygam_calibration=pygam_calibration,
+        split_cal=split_cal,
+        dict_cal_divider=dict_divider,
+        write_library=write_library,
+        use_library=use_library,
+        batch_num=batch_num,
+        n_jobs=n_threads,
+        verbose=verbose,
+    )
 
     # Calibrate the original model based on the new retention times
     if file_cal:
@@ -193,13 +157,16 @@ def run(file_pred,
         preds = dlc.make_preds(seq_df=df_pred, calibrate=False)
 
     df_pred["predicted_tr"] = preds
-    logger.debug("Writing predictions to file: %s", file_pred_out)
+    logger.info("Writing predictions to file: %s", file_pred_out)
     df_pred.to_csv(file_pred_out)
 
     if plot_predictions:
         if file_cal and "tr" in df_pred.columns:
             file_pred_figure = os.path.splitext(file_pred_out)[0] + '.png'
-            logger.debug("Saving scatterplot of predictions to file: %s", file_pred_figure)
+            logger.info(
+                "Saving scatterplot of predictions to file: %s",
+                file_pred_figure
+            )
             plt.figure(figsize=(11.5, 9))
             plt.scatter(df_pred["tr"], df_pred["predicted_tr"], s=3)
             plt.title("DeepLC predictions")
@@ -207,8 +174,10 @@ def run(file_pred,
             plt.ylabel("Predicted retention times")
             plt.savefig(file_pred_figure, dpi=300)
         else:
-            logger.warning('No observed retention time in input data. Cannot \
-plot predictions')
+            logger.warning(
+                "No observed retention time in input data. Cannot plot "
+                "predictions."
+            )
 
     logger.info("DeepLC finished!")